requirements.txt
beautifulsoup4==4.9.3
camelot-py==0.8.2
google-cloud-vision==2.0.0
html5lib==1.1
matplotlib==3.3.2
pandas==1.1.3
pdftotext==2.1.5
Pillow==8.0.0
requests==2.24.0
opencv-python==4.4.0.40
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.0
